autoaim also shoots note once it reaches position

+++ src/main/java/org/first5924/frc2024/subsystems/feeder/FeederIO.java
@@ -13,7 +13,7 @@ public interface FeederIO {
         public double motorTempCelsius = 0.0;
         public double motorCurrentAmps = 0.0;
         public double motorCurrentVelocity = 0.0;
-        public int distanceToNextObject = 0;
+        public double distanceToNextObject = 0;
     }
